pyproject.toml
[tool.poetry]
name = "sqltodf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.1.4"
polars = "^0.19.19"
jaydebeapi = "^1.2.3"
pyodbc = "^5.0.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
pytest = "^7.4.3"
setuptools = "^69.0.2"
pyarrow = "^14.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
